[pr-5] Move DashCardMenuItem to the embedding-sdk plugins contract
It's imported by shared/embedding-sdk but it lived in feature/dashboard,
moved it down to remove violation that will be enforced on one of the next
commits.
