+++ frontend/src/metabase/dashboard/components/DashCard/DashCardMenu/DashCardMenuItems.tsx
@@ -5,13 +5,15 @@ import { canDownloadResults } from "metabase/common/utils/dataset";
 import { editQuestion } from "metabase/dashboard/actions";
 import { useDashboardContext } from "metabase/dashboard/context";
 import { transformSdkQuestion } from "metabase/embedding-sdk/lib/transform-question";
-import type { DashboardCardCustomMenuItem } from "metabase/embedding-sdk/types/plugins";
+import type {
+  DashCardMenuItem,
+  DashboardCardCustomMenuItem,
+} from "metabase/embedding-sdk/types/plugins";
 import { useDispatch } from "metabase/redux";
 import { Icon, Menu } from "metabase/ui";
 import type Question from "metabase-lib/v1/Question";
 import type { DashCardId, Dataset } from "metabase-types/api";
 
-import type { DashCardMenuItem } from "./dashcard-menu";
 import { canEditQuestion } from "./utils";
 
 type DashCardMenuItemsProps = {
