+++ frontend/src/metabase/embedding-sdk/types/plugins.ts
@@ -1,8 +1,10 @@
+/* eslint-disable-next-line no-restricted-imports */
+import type { MantineColor } from "@mantine/core";
 import type { ReactNode } from "react";
 
-import type { DashCardMenuItem } from "metabase/dashboard/components/DashCard/DashCardMenu/dashcard-menu";
 import type { ClickAction } from "metabase/visualizations/types";
 
+import type { IconName } from "./icon";
 import type { MetabaseQuestion } from "./question";
 
 export type MetabaseDataPointObject = {
@@ -35,6 +37,53 @@ export type DashboardCardMenuCustomElement = ({
 }: {
   question: MetabaseQuestion;
 }) => ReactNode;
+
+export type DashCardMenuItem = {
+  /**
+   * Icon name
+   */
+  iconName: IconName;
+
+  /**
+   * Item label
+   **/
+  label: string;
+
+  /**
+   * Item children
+   */
+  children?: ReactNode;
+
+  /**
+   * Key of `theme.colors` or any valid CSS color
+   **/
+  color?: MantineColor;
+
+  /**
+   * Determines whether the menu should be closed when the item is clicked, overrides `closeOnItemClick` prop on the `Menu` component
+   **/
+  closeMenuOnClick?: boolean;
+
+  /**
+   * Section displayed on the left side of the label
+   **/
+  leftSection?: ReactNode;
+
+  /**
+   * Section displayed on the right side of the label
+   **/
+  rightSection?: ReactNode;
+
+  /**
+   * Disables item
+   **/
+  disabled?: boolean;
+
+  /**
+   * Click handler
+   */
+  onClick: () => void;
+};
 
 export type CustomDashboardCardMenuItem = ({
   question,
